Setup test harness to work with MathQuill
Summary:
Eventually we'll want to test React components in the test harness too,
but that will come in a separate diff.  I also, want to wrap MathQuill
so that we have a more uniform API.  This will come in a separate diff
followed by the actual tests to actually test the state transitions for
different actions.

Test Plan: - npm test (see that one test passes)

Reviewers: charlie

Reviewed By: charlie

Differential Revision: https://phabricator.khanacademy.org/D26715