Make text on buttons unselectable
Summary:
This makes it so that you can't select the iconography on the buttons, and the cursor appears as a
pointer. We'll likely want to refine some of this as we do more testing on mobile devices (in fact,
these changes only seem to affect desktop, in my testing--text was never selectable on my phone),
but useful for now.

Test Plan:
Open `index.html` and verify that the buttons now behave as buttons, rather than text fields
(i.e., the cursor is a pointer and you can't grab the text).

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26752