Animate transition between keypad pages
Summary:
This diff adds a transition animation when the user flips between pages in a keypad. In order to
achieve the desired effect, we have to render both pages of keys (one of which is offscreen) and
initiate a CSS animation when the page index changes.

On Asana @ https://app.asana.com/0/114493820703878/117223962064944.

Test Plan:
- Open `index.html`.
- Flip to one of the Expression keypads.
- Verify that clicking '...' opens the second page, and that the page slides in.
- Verify that clicking '123' returns to the first page, and that the page slides in.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26846