Set button heights to be 1/5 of the screen width
Summary:
Here be dragons.

Our keypads are based on a 4x5 grid layout. In the mocks, the keypads are designed such that if you
have a true 5-column keypad, the buttons have a fixed aspect ratio of 1:1. If you have fewer than 5
columns, the buttons maintain the height that they would've had in a 5-column keypad.

With this diff, we dynamically compute the desired height of our buttons and re-render the entire
keypad on resize. This is a poor solution, but thankfully, screens don't resize often (if at all)
in the mobile context.

For testing purposes, I've also implemented a fail-safe whereby we'll truncate the button height if
the target height would make the keypad taller than the screen.

On Asana @ https://app.asana.com/0/114493820703878/116888831022492.

Test Plan:
- Open `index.html`.
- Choose one of the Expression keypads.
- Adjust your screen size (you'll probably have to shrink it down horizontally to trigger the aspect ratio behavior).
    - Verify that the buttons are square and respond to adjustments in screen size.
- Choose a non-Expression keypad.
    - Verify that the button heights match that of the Expression keypad.

Reviewers: emily, kevinb

Reviewed By: emily, kevinb

Subscribers: jared, emily

Differential Revision: https://phabricator.khanacademy.org/D26853