Redesign button configuration props and introduce wrapper to pass them down
Summary:
This diff redesigns our keypad button configuration props. Instead of including the action that a
key should take on press, buttons are now configured with a unique identifier, and the role of
mapping those identifiers to actions will later be handled by the store. In addition, multi-symbol
buttons are now represented as single buttons that compose other buttons by listing their IDs. In
this new system, a key is uniquely identified by its identifier, which is used to figure out what
action it should take on press, figure out what symbol it should render, etc.

To keep `KeypadButton` ignorant of this ID scheme (which will be useful when we implement our
gesture system in a later diff), I've also introduced a `TouchableKeypadButton`, which passes down
sanitized props to the `KeypadButton`. In the future, `TouchableKeypadButton` will use the ID
system combined with other state from the store to tell its child `KeypadButton` to render as
focused, for example.

As of now, all button interactions are disabled.

Test Plan:
- Open each keypad.
- Verify that it renders correctly. (No interactions work after this diff.)

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26957