Wire up focus and popover state from the store to the keypad
Summary:
This diff introduces a `gestureReducer` that will handle the focus and popover state of the keypad,
as well as some more complex gesture state and configuration options in the near future. The gesture
state is wired to the `TouchableKeypadButton`s, which then wire it down to the button that they
manage as simple `focused` and `popoverEnabled` props, keeping the buttons opaque to the ID scheme.

Depends On: D26958

Test Plan:
- Open the `TEST_POPOVER` keypad.
    - Verify that the popover is active, and the top-most button is focused.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26959