Add `GestureManager` to translate touch events
Summary:
This diff adds a thin `GestureManager` that allows for DOM nodes to be linked to particular IDs and
translates raw touch events from the browser into events that signify the nodes over which they
occur.

Right now, the `GestureManager` doesn't do anything with those events--it just logs their
occurrences to the console. In the future, we'll use this to implement complex gestures in our
state machine.

Test Plan:
- Open up any keyboard.
- Open up the console.
- Press on a button.
    - Verify that the correct button is logged as receiving an `onTouchStart` event.
- Drag to another button.
    - Verify that, as your cursor moves over different buttons, they're logged as receiving `onTouchMove` events.
- Release your cursor over a button.
    - Verify that the button over which you released your cursor is logged as receiving an `onTouchEnd` event.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26960