Convert SVG images to React components
Summary:
This diff migrates all of our SVG images over to proper React components, which will allow us to
have a much greater degree of control over how their render and, in particular, allow us to modify
their colors with ease. As an added beneift: all of the icons will come down the JS payload, so the
keypad buttons won't load one-by-one.

In a future diff, I'll modify the highlight state of the buttons to render the SVG text in white.

For reference, here's the [hacky script](https://gist.github.com/crm416/3c7abc88e520eaed72347af240b32590) I'm using to convert the icons.

On Asana @ https://app.asana.com/0/114493820703878/119093262696497.

Test Plan: NFC -- verify that the rendering of the keypad buttons is the same as before.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26994