Inset focus state and echo boxes
Summary:
The focus state for the keypad buttons is meant to be inset slightly, and the echo animations are
only supposed to be as large as this focus state. This diff implements that behavior by rendering
the focus background as a separate `<View>` and enforcing a transparent background on the primary
`<View>` when rendering an echo.

Test Plan:
- Open up a keypad.
- Touch down on some buttons.
    - Verify that the focus state is evenly inset.
    - Verify that the focus state has a 1px border radius.
- Click on some buttons.
    - Verify that the echo matches the size of the focus state.
    - Verify that the echo has no white background outside of the green focus state (i.e., it doesn't obscure anything beneath it in those regions).

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27016