Fix sidebar in place on multi-page keypads
Summary:
For the multi-page keypads (in which you can swipe and change pages), the sidebar (which houses all
of the control keys, like dismiss and delete) is meant to remain fixed in place while the rest of
the keypad swipes back and forth. This diff implements that behavior by introducing the concept of
a `sidebar` into our `TwoPageKeypad` component.

On Asana @ https://app.asana.com/0/114493820703878/119110690483561.

Test Plan:
- For both the Basic and Advanced Expression keypads:
    - Verify that the sidebar takes up 1/5 of the screen, on all pages.
    - Verify that the sidebar remains fixed when changing pages (using the top-right button).
    - Verify that the top-right button toggles between '...' and '123' as you change pages.

In GIF form: {F518051}

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27029