Avoid unnecessary coordinate-to-node computations
Summary:
Often, the `GestureStateMachine` doesn't need the identifier of the DOM node over which a touch
event occurred, yet we send it anyway. For example, when the user is swiping, we never end up
using the node identifier, but we're computing it on every move event.

This diff converts the `GestureStateMachine` to take a function that computes a node identifier,
and requires it to call that function when the identifier is actually needed.

In addition, there are two more optimizations related to DOM node lookups:

1. When registering a touch-start event, we can use the ID of the node that received the event, rather than look it up.
2. When registering a touch-end event, we can use the ID of that node that is currently focused, as long as there hasn't been a long-press on that node (in which case, the node that was last focused may not be the node that should receive the touch, as a popover may have appeared on top of it without the user moving their finger).

Further wins could be had by merging the `PopoverStateMachine` with the `GestureStateMachine`, although the
optimizations here will handle the most performant-dependent cases (i.e., quick taps).

On Asana @ https://app.asana.com/0/114493820703878/131197441810923.

Test Plan:
- Run a trace in DevTools and verify that `idForCoords` is not making a big dent in the scripting time. Further, there should no longer be any layout happening, since we're no longer calling `elementFromPoint`, which triggers a reflow.
- Run through the gamut of gestures:
    - Clicking on a popover node.
    - Swiping before a popover node reveals its popover.
    - Clicking on a popover node, waiting for its popover to appear, then releasing.
    - Releasing off a popover node after its popover has appeared.
    - Dragging around vertically on non-popover nodes.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27227