Avoid unnecessary coordinate-to-node computations
Summary:
Often, the `GestureStateMachine` doesn't need the identifier of the DOM node over which a touch
event occurred, yet we send it anyway. For example, when the user is swiping, we never end up
using the node identifier, but we're computing it on every move event.

This diff converts the `GestureStateMachine` to take a function that computes a node identifier,
and requires it to call that function when the identifier is actually needed.

Test Plan: NFC. Run a trace in DevTools and verify that `idForCoords` is not making a big dent in the scripting time. Further, there should no longer be any layout happening, since we're no longer calling `elementFromPoint`, which triggers a reflow.

Reviewers: kevinb