Add `shouldComponentUpdate` guard to prevent buttons from re-rendering
Summary:
To improve swipe performance, this diff adds a `shouldComponentUpdate` guard to our
`<TouchableKeypadButton/>`. The check can be relatively simplistic, as much of our system makes
strong assumptions about the uniqueness of IDs, which makes it easy to tell when a button needs to
update.

On Asana @ https://app.asana.com/0/114493820703878/119470022939187.

Test Plan:
- Add some `console.log` statements to the `render()` methods of `<TouchableKeypadButton/>` and `<KeypadButton/>`.
- Verify that there are no renders as you swipe (previously, every button was re-rendering on every move event).

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27228