Explicitly support a single key handler
Summary:
Previously, we had some logic for managing multiple key handlers. This diff removes that logic in
favor of a single key handler, which is what we will likely want long-term and will make it easier
to use the result of the `handler` function.

Test Plan: Open up a keypad, type some math, and verify that the keystrokes are recognized.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27136