Have the cursor handle track the finger
Summary:
In a future diff I'm going to add in some friction so that it's harder
to drag the handle outside of the math input area.  I also want to switch
to SVG to make it easier to do the necessary calculations.  This diff
causes the cursor handle to track the fingure exactly everywhere.  When
the user touches up, the handle will snap into place under the cursor.
In a future diff this will be animated.

Test Plan:
- drag the cursor around in some math that contains fractions
- it should snap to the correct place
- it should move the cursor correctly when dragging

Reviewers: charlie

Reviewed By: charlie

Differential Revision: https://phabricator.khanacademy.org/D27142