Update button height on initial render
Summary:
Previously, we were only updating the button height when the page resized, and not on the initial
render. That meant that to get the buttons to size correctly, you _had_ to resize the page.

This diff adds an additional update to the button height that occurs after the initial render.

On Asana @ https://app.asana.com/0/114493820703878/130234333349700.

Test Plan:
- Open up a keypad
- Verify that the buttons are correctly sized on first render (square for the Expression keypads).

Initial render before:

{F525002}

And after:

{F525003}

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27155