Tweak echo animation borders to exactly match the focus state of the buttons
Summary:
The echo animations were off by a few pixels in the horizontal and vertical directions, depending on
where the button was in the keypad. This was a result of:

1. The top border on the keypad.
2. The borders on the keys themselves, which varied from button to button.

As far as I can tell, the root cause is that `getClientBoundingRect()` includes borders in its
computation, which means that they need to be accounted for when we're repositioning based on the
box that it returns. Unfortunately, as our buttons have variable borders, this means that our echo
states also need to be aware of the borders on the buttons over which they'll appear. This isn't
challenging to do given our existing infrastructure, but it does mean that the gesture system is
now 'aware' of the idea of borders.

On Asana @ https://app.asana.com/0/114493820703878/125597594329566.

Test Plan:
- Open up any keypad, with every variant of the pagination UI (since they have some effect on padding for the keypad).
- Click each button.
    - Verify that there is no offset between the echo animation and the focus state.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27181