Disallow cursor placement within commands
Summary:
MathQuill doesn't treat commands (like \cos and \sin) as atomic units; instead, it treats them as
lists of characters. However, we don't want users to be able to edit these commands as individual
characters, as they'll only be entering them by pressing a single button.

This diff disallows the cursor from being placed within the body of a command, including its parens.

The basic approach is to detect if the node at which the cursor is being placed is part of a command
and, if so, to place it at the start of the interior of its parens.

On Asana @ https://app.asana.com/0/114493820703878/130801370525428.

Test Plan:
- Run the unit tests (as the "delete the command if backspacing from the start of its parens" behavior is now implemented as a special case of this behavior).
- Open up a keypad.
- Type in some math, followed by 'cos(45)', for example.
- Drag the cursor handle.
    - Verify that whenever you're over anything in 'cos(', the cursor is snapped to the inside of the parens.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27188