Activate keypad when input receives a touch event
Summary:
Now, whenever the user touches on the math input box, the keypad will reappear. This will mostly
be useful for user testing, so that the keypad isn't eternally dismissed as, in the future, we'll
likely re-activate the keypad in response to the input being refocused by Perseus, rather than in
response to a direct touch event.

Test Plan:
- Open up the keypad.
- Dismiss it.
- Tap on the input.
    - Verify that it reappears.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27195