Add support for theta and require all variables and constants to be capitalized
Summary:
This makes the contract for specifying `extraSymbols` a little clearer for clients. Rather than
worrying about the need to capitalize `PI` but keep variables lower-case, all variables and symbols
are now required to be capitalized.

(We could also perform this capitalization internally, but that seems a little too 'magical' from a
client's perspective.)

In addition, I've added in `THETA`, as it appears in some of the mocks. I expect that we'll need to
add support for a few more Greek symbols as well, so there's a TODO in here.

Test Plan:
- Set the Advanced keypad to use `PI`, `THETA`, and some non-Greek vars in `keypads.js`.
- Open up the Advanced keypad.
    - Verify that the keys render as expected.
    - Verify that each key triggers its symbol to be written in the input.
        - Verify, in particular, that the variables are lower-cased when they appear in the input.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27272