Avoid re-render before `onChange` has propagated
Summary:
As `onChange` is asychronous, and we're immediately calling `setState` afterwards, the cursor would
often disappear as the contents of the math input would be squashed in `componentDidUpdate`, since
`this.props.value` wouldn't be updated in time for the render following the `setState` call, and so
the component would act as if the value was being overridden from the outside.

Instead, we need to do the `setState` as a callback to `onChange`, which is something that Perseus
supports.

Test Plan:
- Modify `onChange` to take some amount of time; e.g., wrap it in a `setTimeout`.
- Type some math.
    - Verify that the cursor is visible as you type and the math updates properly.
- Test in Pereus, too.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27311