Scroll focused input into view
Summary:
As the keypad is quite large, we need to scroll the active input to the top of the page when the
user focuses it, so that it's not obscured by the keypad. This diff implements that logic,
drawing its implementation from `phone-help-bar.jsx` in `webapp`'s `exercises-package`.

On Asana @ https://app.asana.com/0/114493820703878/133274501859654.

Test Plan:
- Change `index.html` to have `min-height: 1000px`.
- Change `marginTop` on the input wrapper in `app.js` to something large (like `300px`).
- Tap on the input.
- Verify that it appears scrolled to the top, with some margin.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27345