Explicitly set background color on input
Summary:
Otherwise, in some cases (like when the input is included as part of a table), it inherits a
background color.

Test Plan: Open up an exercise in webapp with inputs in a table ([link](http://localhost:8080/math/cc-fourth-grade-math/cc-4th-mult-div-topic/cc-4th-area-models/e/multiplying-2-digit-numbers-with-area-models)); verify that they have white backgrounds, not the background of the table cells.

Auditors: kevinb