Fix popover shadow and animation styling
Summary:
This diff refactors the popover rendering so as to fix the box shadow and add in fade-in and
fade-out animations. The change is fairly hairy, as the only way to prevent the box shadow from
being cutoff by the sidebar is to render the popovers outside of their actual keys (since the
keys themselves are in a different stacking context than the sidebar due to the use of `transform`
for the multi-page slide effect); as such, the manner in which we render the popovers now resembles
the manner in which we render the echoes (i.e., their absolute positions on the page are computed,
and then mapped to relative positions within the keypad before being rendered directly on top of
it, independently from the keys over which they're meant to appear).

On Asana @ https://app.asana.com/0/133656453415343/121486514794981 & https://app.asana.com/0/133656453415343/131139018244294.

Test Plan:
- Open up `index.html`.
- Choose the 'Basic Expression' or 'Advanced Expression' keypad.
- Tap on a multi-function key.
    - Verify that the popover renders in the same position as before.
    - Verify that the box shadow is not cutoff by the sidebar.
- Drag up and down the popover.
    - Verify that the focus follow your finger.
- Release on a node in the popover.
    - Verify that the proper key is triggered.

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27382