Fix three-variable case
Summary:
It turns out that the case in which the keypad needs to present exactly three variables was causing
a JS error. This diff fixes it by only rendering the icon for the fourth variable if it is present
(though it still renders a placeholder view there so-as to keep the four-quadrant layout intact).

Test Plan:
- Open up the keypad.
- Choose the Basic Expression keypad (which I've configured with three symbols).
    - Verify that they render in a four-quadrant layout, but that only three symbols are present.
    - Verify that there are no console errors.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27737