Hide corner decals on focus
Summary:
Our multi-functional keys display decals in the corner to indicate that they can be long-pressed.
However, we want to hide these decals on focus. This change was slightly more involved than the
one-liner that I anticipated due to the animation of the popover state, which requires that we not
only hide the corner decal on focus, but also when the key is the parent of a popover (to which the
keys used to be ignorant).

On Asana @ https://app.asana.com/0/132409062806753/139327178321428.

Test Plan:
- Open up the Basic or Advanced Expression keypad in the demo page.
    - Verify that the corner decals are present.
- Touch down on a multi-functional key.
    - Verify that the corner decal disappears immediately, and stays gone even as the modal animates into view.
- Release.
    - Verify that the corner decal re-appears.

{F542355}

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D27776