Allow clients to provide custom styles for math input
Summary:
This diff adds a `style` prop to the `MathInput` component, so as to allow clients to provide
custom styles. This will be used, in particular, for the `Matrix` widget in Perseus, which needs
to make some minor alterations to the background color and do some custom re-sizing.

Test Plan:
- Add a style prop to the `MathInput` rendered in `app.js`.
- Open up the keypad.
    - Verify that the style props are respected.
    - Verify that all the existing styling and MathQuill interactions are preserved.

Reviewers: david

Reviewed By: david

Subscribers: shadaj, benkomalo

Differential Revision: https://phabricator.khanacademy.org/D27871