Re-enable cursor context tracking
Summary:
This diff re-enables the cursor context tracking behavior that allows the input to relay cursor
context changes to the keypad (i.e., in the event that the cursor context changes outside of a
keypress, as when the user drags the cursor themselves).

Unfortunately, this change will result in even more coupling between the keypad and the input;
however, in my assessment, this is no worse than calling `this.props.keypadElement.setKeyHandler`.

Test Plan:
- Open up the keypad in the demo, with dynamic cursor tracking enabled (`jump_out_type: dynamic` in `custom.html`).
- Type some math.
    - Verify that the 'Jump Out' button swaps with the 'Dismiss' button (bottom-left corner) based on the level of nesting, both:
        - As you type.
        - As you drag the cursor.

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: shadaj

Differential Revision: https://phabricator.khanacademy.org/D27875