Pre-inject echo animation styles
Summary:
This fixes an oft-noticed bug in which the echo views change in size during the lifecycle of their
animation. Similar to D27867, the issue seems to be related to flushing styles in Safari.
Unfortunately, the fix in this case is a bit more involved, since we have a greater number of
permutations in the style parameters.

Test Plan:
- Open up the keypad in mobile Safari.
- Tap on a bunch of keys.
    - Verify that there is no inconsistency in the width of the echoes (i.e., that they don't change in size).
- For an additional guarantee:
    - Add some logging to Aphrodite, to log when it injects a new style.
    - Repeat the test plan from above.
        - Verify that there are no style injections when tapping on an input.

Reviewers: benkomalo, emily

Reviewed By: emily

Subscribers: abdulrahman

Differential Revision: https://phabricator.khanacademy.org/D27971