Disable swiping when starting from the navigation bar
Summary: Right now, it is possible to initiate a swipe by pressing on a key in the navigation bar. Since the navigation bar does not move with a swipe, it can be confusing to swipe the navigation bar but see other elements move instead. Now, it is not possible to initiate a swipe from the navigation bar, and so swipes must start from the main key panel instead.

Test Plan:
- Verified in a browser that it is still possible to swipe from non-navigation keys
- Checked that it is still possible to use all the keys on the keypad
- Checked that it is still possible to move over the navigation bar during a swipe

Reviewers: charlie, kevinb

Reviewed By: charlie, kevinb

Differential Revision: https://phabricator.khanacademy.org/D28012