Provide support for customizing fraction behavior on a per-input basis
Summary:
This diff adds support for customization of the fraction behavior on a per-`KeypadInput` basis.
Previously, the behavior was only configurable globally through the `Settings` module.

In the Perseus substate, we'll select the appropriate fraction behavior based on the keypad
configuration parameters for the input.

On Asana @ https://app.asana.com/0/132409062806753/130801370525433.

Test Plan: Same as D27499.

Reviewers: kevinb, shadaj, benkomalo

Reviewed By: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D28111