Revert `overflowX` change
Summary:
The use of `overflowX: 'hidden'` was causing the popover menus and echo states to be cutoff at the
top of the keypad, for unknown reasons. We should resolve this in the near future, but for now, it
seems prudent to disable that style and accept that the extra pages may be visible in some edge
states.

Test Plan:
- Substate this change into Perseus.
- Verify that the echo states and popover menus are not cutoff in any way.

Auditors: benkomalo, shadaj

Subscribers: kevinb