Add README and LICENSE
Summary:
Now that the keypad has matured a bit, it seems wise to add a small README outlining some of its
behaviors, as well as a license to match our other open-source projects.

This could use a lot more work, but then again, there's also some clean-up work that we'd need to
do to make it easily usable and customizable in a general context.

Test Plan: None!

Auditors: kevinb, benkomalo