Force visibility of the cursor tapping mid-selection
Summary:
Previously, there were certain cases where, when tapping on an input with an active selection, we'd
hit a null error. The issue was that we weren't _guaranteeing_ that the cursor would be active by
`onTouchEnd`, which assumes an active cursor. A sample failure case would be if you wrote "9/9",
hit the right arrow, then hit delete and tapped just to the right of the selection area. Then,
MathQuill wouldn't force the cursor visibility on its own, since your tap would be slightly outside
of MathQuill's domain, and we wouldn't force it in `focus()`, since the input is already focused.

On Asana @ https://app.asana.com/0/132409062806753/143008249310754.

Test Plan:
- Open up the keypad.
- Type "9/9", then hit "Right", then "Backspace".
- Tap to the right of the selection area.
    - Verify that the cursor appears on touch down, and there is no error.

Auditors: kevinb