Insert an empty set of parens when adding exponent at invalid location
Summary:
Previously, we allowed the user to insert an exponent after an expression that couldn't be raised
to a power, and at the beginning of the input, when there was no content preceding it. This diff
modifies the math input behavior to instead insert an empty set of parentheses when the exponent
would otherwise be invalid.

For the case in which we insert a square or a cube with an empty set of parentheses, we also move
the cursor into the parens.

On Asana @ https://app.asana.com/0/132409062806753/136371284656394.

Test Plan:
- Run `npm test`.
- Add a bunch of operators to the input (any arithmetic or equality operators).
- Add a custom exponent, square root, and cube root after each operator.
    - Verify that, in each case, a set of empty parentheses is added after the operator and before the exponent.
    - Verify that, when a square root or cube root is added with an empty set of parentheses, the cursor ends up in the parentheses.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D28259