Remove margin from cursor to prevent the cursor from changing the height of its parent
Summary: Previously, moving the cursor around would cause elements to shift because the cursor would change the height of its parent. Now, the margin from the cursor has been removed so that the height is unchanged.

Test Plan: - Verified in a browser that moving the cursor around no longer causes elements to shift

Reviewers: charlie, kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D28272