Add multi-touch tracking to gesture state machinery
Summary:
This diff adds multi-touch tracking to the `GestureStateMachine` class, along with a suite of unit
tests to validate the new behavior.

The basic semantics of the system are such that we now track most of the gesture state on a
per-touch basis, so that, e.g., we only trigger swipes if a given _finger_ moves past the swipe
displacement threshold (rather than if the user placed their finger down on the left edge, then
placed a different finger down on the right edge of the keypad). Furthermore, I've opted to ignore
all existing and new touches (apart from the touch that is controlling the swipe) as soon as a
swipe is triggered.

There are still a few weirdnesses or inconsistencies that fall out of the `GestureStateMachine`
operating on a multi-touch basis, but the `PopoverStateMachine` remaining monolithic. (For example,
tapping to reveal a popover and then releasing with a separate finger would cause the popover to
dismiss.) These can be ironed out later.

Test Plan:
- Run the new unit tests (`npm test`).
- Open the keypad on a physical device and verify that:
    - Single-finger interactions remain as before: tapping, dragging, long-pressing, and holding are all NFC in the context of a single finger.
    - Tapping and holding a popover and then releasing without ever moving your finger triggers the default node.
    - Touching down on a key on the left side of the keypad and then touching down on a key on the right side of the keypad does not trigger a swipe.
    - Any fingers--apart from the swiping finger--are completely ignored as soon as a swipe starts.

Reviewers: kevinb, jared

Reviewed By: jared

Differential Revision: https://phabricator.khanacademy.org/D28353