Enforce a max-width on the keypad and center
Summary:
https://app.asana.com/0/145577216935888/146563261101666

This is a stopgap fix for keypads on tablets, since we're now
enabling them on tablets in addition to phones. Before we sink
a bunch of time optimizing the layout for larger screens, we're
putting in a stopgap fix so that the 1:1 aspect ratio requirement
doesn't make the keypad take up an obscene amount of space.

This change makes the max-width of the keypad to be 512, and
centers the content. There should be no user-visible differences
on mobile devices (below 512)

Test Plan:
 - pass the `web_tablet_exercises` gandalf bridge
 - using [[ http://localhost:8080/math/integral-calculus/solid_revolution_topic/volume-solids-known-cross-sectio/e/volumes-of-solids-of-known-cross-section | an exercise with math input ]]...
 - load it on iPad chrome emulation:
 - slide keypad back and forth
 {F545073}
 - load it on iPhone5 chrome emulation:
 - slide keypad back and forth
 {F545074}

Reviewers: david, charlie

Reviewed By: charlie

Differential Revision: https://phabricator.khanacademy.org/D28550