Fix spacing violations picked up by new lint rules
Summary:
Before the spacing rules added to our linter in D28403 are turned on,
I need to grandfather in or fix up code with existing violations.
Since this is a subrepo of Perseus, the linter was complaining.

Rather than adding a bunch of noisy lint-disable lines, I opted to let
the auto-fixer have a go. The only violation was
`object-curly-spacing`, so `arc lint` did the trick.

Test Plan: `ka-lint`

Reviewers: kevinb, charlie

Reviewed By: charlie

Differential Revision: https://phabricator.khanacademy.org/D28931