Remove sidebar layout
Summary:
With the new layout for the Expression keypad, we no longer have the concept of a fixed sidebar
that persists across multiple pages of keys. (Instead, the keys that were previously part of the
sidebar will now be part of the first page, and inaccessible from the second page.) This diff
removes the sidebar layout--for now, I'm just replicating the sidebar as an additional column in
both pages, though it will be removed in the next diff.

On Asana @ https://app.asana.com/0/157213482411363/157213482411370.

Test Plan:
- Open up the Expression keypad.
    - Verify that the sidebar appears in both pages.
    - Verify that you can initiate a swipe from the sidebar.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D29116