Create a separate 'negative' key
Summary:
In the context of the Fraction keypad, the 'minus' key is really a 'negative' key and thus needs to
be rendered as a 'value' key ratehr than as an 'operator'.

On Asana @ https://app.asana.com/0/157213482411363/157333652496580.

Test Plan: Open up the Fraction keypad; verify that the minus key is rendered in white and works as before.

Auditors: kevinb