Fix borders between keypad pages and navigation pad
Summary:
We need to include hairline borders between any two pages of keys, as well as between the keypad
and the navigation pad, with the additional constraint that the borders between the pages of keys
in the view pager should never be visible when the pager is stationary (i.e., the borders need to
be rendered between the pages). This brings our borders up to spec in all of those cases!

The view pager logic has also become substantially simpler, since we can now assume that a page is
the width of the screen. (Before, our pages only consisted of the first four columns, given the
fixed sidebar. This required measuring the pages and messing about with a lot of component
nesting.)

On Asana @ https://app.asana.com/0/157213482411363/158470495957511.

Test Plan:
- Open up the Expression keypad on a portrait iPhone.
    - Verify that the visible page is 320px wide.
    - Verify that, when swiping left, a border becomes visible between the pages.
    - Verify that, after changing pages, there's no visible border between the pages.
- Rotate to landscape.
    - Verify that there's a 1px border between the two pages.
    - Verify that both pages have the same width (so, the border is excluded when calculating the widths).
- Transition to an iPad.
    - Verify that, for all keypads, in all orientations, there's a 1px border between the navigation pad and the keypad.

{F548177}

Reviewers: benkomalo, kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D29220