Avoid blurring input when touches occur below keypad
Summary:
This is a hack to avoid dismissing the keypad when the user presses the 'Check answer' button. Now,
instead of merely checking if a given touch occurred within the keypad, we also permit touches
below the keypad (such that if the keypad is anchored above the bottom bar, the user can still
check their answer without dismissing the keypad).

Test Plan: Coming in the integration diff. For now, verify that the keypad dismisses on touch-outside, and remains active when pressing keys.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D29322