Increase z-index of handle
Summary:
The handle is below the 'Stuck? Get some help' text on the web, which has a z-index of 3
(`interactiveComponent). This is a quick, hacky fix to move it above that copy.

Test Plan: Open in webapp; tap on the input; verify that the cursor is above the 'Stuck?' text.

Auditors: kevinb