Blur when touches occur to the left or right of keypad
Summary:
This diff makes two minor modifications to fix the 'blur on touch behavior'. First, it passes up
the node that corresponds to the 'compact' or 'fullscreen' keypad, rather than its parent (which
is always fullscreen). And second, it removes the requirement that touches that occur below the
keypad also occur within it horizontally (which was a bug).

Test Plan:
- Substate into webapp.
- With a full-width keypad, verify that tapping within the keypad or on the bottom bar doesn't dismiss, but anywhere else does.
- With a compact keypad (tablet, fraction), verify that tapping within the keypad or on the bottom bar doesn't dismiss, but anywhere else does.
    - In particular, verify that tapping to the side of the keypad dismisses it, but only when tapping above the bottom bar.

Reviewers: benkomalo

Reviewed By: benkomalo

Differential Revision: https://phabricator.khanacademy.org/D29432