Refine behavior for deleting from within a radical
Summary:
Previously, when deleting from within a radical, if the cursor was flush to the left edge, we'd
leave the radical root intact and only delete the body. This made for a weird experience, where
hitting the 'Cube root' key and then the 'delete' key would leave a '3' alone in the input.

Now, the behavior is as follows:

- When deleting with your cursor on the left edge of a radical body, if the root is non-empty,
  then the entire radical is selected (similar to parentheses).
- When deleting with your cursor on the left edge of a radical body, if the root is empty,
  then the entire radical is deleted (similar to parentheses).

On Asana @ https://app.asana.com/0/132409062806753/144285431869204.

Test Plan:
- Enter some math.
- Enter a cube-root.
- Press 'delete'.
    - Verify that the root is selected but not deleted.
- Move to delete the cube, leaving the root empty.
- Move your cursor into the root.
- Press 'delete.'
    - Verify that the root is deleted.
- Verify that deletion works normally when the cursor is not at the left edge of the radical.
- Verify that deletion works as-before for square roots.
- Run the new unit tests.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D29682