Avoid updating cursor after blurring during multi-touch gestures
Summary:
Some of the gesture-handling code in `math-input.js` assumes single-touch gestures. This was
causing two different reproducible bugs, one of which led to awkward UI behavior and another of
which threw a JS error (which would lead to exercise kick-outs in the native apps). The specific
bugs are described in the test plan.

On Sentry @ https://sentry.io/khanacademyorg/android-webview/issues/194564556/.

On Asana @ https://app.asana.com/0/231953500557188/232852627504033.

Test Plan:
- Open up the keypad on a real device.
- Tap to add some math.
- Touch down in the input. Keep your finger down and drag.
- With a different finger, tap outside the input.
- Release the first finger.
    - Verify that there is no JS error.
    - (Previously, we'd attempt to move the cursor in the touch-end, thereby assuming that the
       cursor was visible, causing a JS error.)
- Repeat, but after tapping outside the input, move the initial finger around instead of releasing.
    - Verify that the cursor doesn't appear.
    - (Previously, the cursor would appear, despite the fact that the input was no longer focused.
       So you'd see a borderless input with a cursor in the middle.)

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: benkomalo, #android

Differential Revision: https://phabricator.khanacademy.org/D32775