Prevent external styling of math in keypad keys
Summary:
Perseus sets the text color of all `.katex` elements to `gray17`. This clashes with the way that
we try to color `.katex` elements rendered in the keypad (by way of setting the `color` on the
parent element, to a different gray - or white, if focused). On web, there was no issue because of
the way that the parent elements are architectured (the keypad isn't a child of the greater Perseus
container, and so the Perseus styles don't apply); on mobile, however, keys containing TeX would
always render as the Perseus-math gray, even when pressed. As a result, we had gray-on-green keys
for the focus state!

On Asana @ https://app.asana.com/0/257584043292808/281398689331264

Test Plan:
- Substate into webapp.
- Run the app.
- Point to webapp.
- Open an exercise with variable keys (e.g., "Distributive property with variables").
- Tap on the TeX key in the bottom left.
    - Verify that it transitions to white!

After: {F631541} (previously, the "y" remained black)

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D34006