Migrate to ES6 classes from React.createClass
Summary:
This diff migrates the math-input React components to use ES6 classes over `React.createClass`.

The heavy-lifting was done with the `class.js` transform from [react-codemod](https://github.com/reactjs/react-codemod), with a bit of hand-tweaking.

Note that the `class.js` transform is conservative in that it transforms all non-lifecycle methods
to property initializers in order to preserve binding, which isn't ideal, but shouldn't be a big
issue.

Test Plan:
- Run `npm run watch`.
- Open up the demo.
- Tap on the input.
    - Verify that the keypad appears.
- Verify that you can swipe.
- Verify that you can enter keys by tapping.

Reviewers: jared, benkomalo

Reviewed By: jared, benkomalo

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D30187