Add `getDOMNode()` to the keypad API to allow for proxying
Summary:
On mobile native, we want to have the keypad be on the native side
instead of in the webview. This means that `keypadElement` won't be a
react element, and `ReactDOM.findDOMNode(keypadElement)` will crash.
This adds `getDOMNode()`, which the proxy can just return `null` from.

This will be followed by a diff to webapp, changing perseus to switch over
`findDOMNode` calls, and adding an apiOption for the native side to register a
keypad proxy.

Test Plan: :fingers-crossed:

Reviewers: kevinb, hannah

Reviewed By: hannah

Differential Revision: https://phabricator.khanacademy.org/D41713