Fix mobile keypad not scrolling out of view in Chrome browser
Summary:
A year ago there was a change to Chrome that makes it return 0 for `document.body.scrollTop`
https://bugs.chromium.org/p/chromium/issues/detail?id=766938

This degraded mobile interaction experience on Android mobile web (and native app)
as the keypad does not scroll out of view when user interacts with math input widget,
potentially obscuring the input.

{F1333462}

This change fixes that by scrolling the appropriate page element (in all browsers).
I have confirmed that this change has no negative effect on iOS Safari (`document.body === document.scrollingElement`).

{F1333463}

I'm not sure as to the steps required to make this fix be part of the Android native app.
Will it be done automatically once the built `math-input.js` is deployed to `webapp`
and becomes part of the Perseus bundle?

https://khanacademy.atlassian.net/browse/CP-497

Test Plan:
 - Run `node_modules/webpack/bin/webpack.js --config webpack.config.externals.js && cp build/math-input.js ../webapp/third_party/javascript-khansrc/math-input/`
 - In webapp run `DEV_APPSERVER_FLAGS='--enable_host_checking=false' make serve`
 - Open `/math/arithmetic/arith-decimals/arith-review-decimals-to-fractions/e/fraction-decimal-intuition` using https://browserstack.com on Android Galaxy device Chrome browser
 - Verify tapping on `numeric-widget` will open keypad and scroll page so that input is not obscured by the keypad

Reviewers: davidflanagan, jared, hannah

Reviewed By: davidflanagan

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D48794