Changes to import prop-types as a module instead of using React.PropTypes
Summary:
This diff enables us to migrate to webpack which relies on react 16 that does not contain PropTypes in it, so Im adding an explicit import for it.

Note on tests.  Tests dont fully pass because they timeout in mathquill.  But whatever tests were passing before this changes are still passing now.

Screenshots:

{F1371199}

{F1371200}

{F1371216}

Test Plan:
1. Run `npm run test`

You can also verify this works in webapp in the webpack branch
1. Start local server `make serve`
2. Enable `use_webpack` gandalf bridge and visit `http://localhost:8080/devadmin/content/exercises/addition_2/3002/x2c34ec3e706a977e/preview`
3. Verify the page loads correctly.
4. Disable `use_webpack` gandalf bridge and visit `http://localhost:8080/devadmin/content/exercises/addition_2/3002/x2c34ec3e706a977e/preview`
5. Verify the page loads correctly

Reviewers: kevinb, john

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D50312