Fix a null reference error
There are some times when the node is not in the list anymore.  In this case, it fails to add an entry to the cache and retrys over and over again, littering the console with errors

Handle this case by checking to see if the node is in the list.  If not, but a stub in place