New cursor handle icon #46
SUMMARY
We want to update the cursor handle (the teardrop icon) to be a little more modern looking with a white outline and a subtle drop shadow. This commit mostly just replaces the svg code in `cursor-handle.js` with the new svg code, but there are also some size variables and styling that I had to update because the size is just a little different now.

One unintended side effect I noticed when testing is that when you pull the cursor above and in the horizontal center of the input field, it now snaps to one side of the input values, not in the center of the input values. So that's a little buggy now. I'm not sure how important that is - if that's a blocking change, I'll dig in and figure out where that's coming from.

I also snuck in another change to the blinker cursor so it's now rounded instead of rectangular.

Jira: https://khanacademy.atlassian.net/browse/LP-7670
Design: https://www.figma.com/file/2lUPOSbOP8tbW7RLqbBFLh/Expression-Widget?node-id=542%3A1369

SCREENSHOTS
Cursor blinker, no handle
BEFORE
<img width="643" alt="cursor_before" src="https://user-images.githubusercontent.com/7761701/73204266-845adf80-40f3-11ea-9739-81c3dbfe37db.png">

AFTER
<img width="636" alt="Screen Shot 2020-01-27 at 10 50 50 AM" src="https://user-images.githubusercontent.com/7761701/73204284-8cb31a80-40f3-11ea-9678-00af56da5b42.png">

Cursor blinker and handle
BEFORE
<img width="637" alt="handle_before" src="https://user-images.githubusercontent.com/7761701/73204364-ad7b7000-40f3-11ea-88fc-1e7c3d5199e2.png">

AFTER
<img width="633" alt="Screen Shot 2020-01-27 at 10 51 15 AM" src="https://user-images.githubusercontent.com/7761701/73204382-b1a78d80-40f3-11ea-92ae-ba21a17deaf1.png">

Gif of pulling the handle around
BEFORE
![before](https://user-images.githubusercontent.com/7761701/73204394-b4a27e00-40f3-11ea-86ca-48cdc61e3ca9.gif)

AFTER
![after](https://user-images.githubusercontent.com/7761701/73204398-b9673200-40f3-11ea-9c1f-5f1e29396680.gif)

TEST PLAN
Open the index.html file,
open the dev tools and turn on mobile simulation,
verify the unfocused state looks the same,
tap on the input field,
verify the blinking cursor appears and is now rounded.
Click a buncha buttons on the keypad,
tap into the input field,
verify the cursor handle appears and it looks the same but now has a white outline and a slight drop shadow,
pull it around,
and verify it works as expected.