Give the input field scroll bar more space
SUMMARY
Issue: https://khanacademy.atlassian.net/browse/LP-7673
Design: https://www.figma.com/file/2lUPOSbOP8tbW7RLqbBFLh/Expression-Widget?node-id=542%3A1369

In PR #48 https://github.com/Khan/math-input/pull/48 I added scrolling functionality to the expression widget input box. There are a couple of niggling things that bug me about it, and one of those things is that the scroll bar is super small, hard to see, and hard to grab onto. The issue is that the scrollable element is not very tall; all the padding in the input field belongs to its parent.

This diff moves the padding down one level to the scrollable element. That gives the scroll bar the space it needs while also not interfering with the logic around how much padding to put above and below the text.

There are a couple of improvements that I'm planning to do:
- scroll the input field to the right when new values are added after the max width of the input field has been reached (I need to make a Jira for this)
- scroll the input field if the values have exceeded its max width and the user pulls the cursor handle off to the side where values are being hidden (Jira: https://khanacademy.atlassian.net/browse/LP-7695 )

TEST PLAN
Navigate to /khan/math-input/index.html with mobile simulation on,
enter a few "normal" values (numbers and simple operators),
verify that the text is vertically centered,
add some more complicated values (like fractions and square roots),
verify that the text is still vertically centered,
enter in a whole bunch of values so you exceed the max width of the input field,
and verify the input box is horizontally scrollable.
Click/tap into the input field,
verify the cursor handle appears and looks good,
grab the handle,
pull it and let go in every direction you can think of,
and verify the handle + blinky cursor snap back to the input field at the location you would expect based on where you let go.