Add hover and active state styling [WIP]
Jira: https://khanacademy.atlassian.net/browse/LP-7737
Design: https://www.figma.com/file/2lUPOSbOP8tbW7RLqbBFLh/Expression-Widget?node-id=542%3A1369

This is a WIP that I'm not going to finish right now because I can't test the hover work yet. But I'm putting it out in a PR for posterity so it's easier to pick up later. The big thing that still needs done is implementing the style changes to the `_root` element in the `handleMouseEnter` and `handleMouseLeave` methods.